Version Packages (#118)
This PR was opened by the [Changesets
release](https://github.com/changesets/action) GitHub action. When
you're ready to do a release, you can merge this and the packages will
be published to npm automatically. If you're not ready to do a release
yet, that's fine, whenever you add more changesets to main, this PR will
be updated.


# Releases
## @cloudflare/workers-oauth-provider@0.1.1

### Patch Changes

- [#114](https://github.com/cloudflare/workers-oauth-provider/pull/114)
[`768cd6c`](https://github.com/cloudflare/workers-oauth-provider/commit/768cd6c9d34488f653a678b08f33070b31c071e5)
Thanks [@DeanMauro](https://github.com/DeanMauro)! - adds a method
`decodeToken` that retrieves a granted access token from the KV and
returns the user-defined props attached to it. This permits token
decoding outside of a fetch call, e.g. an RPC call from another worker.

Co-authored-by: github-actions[bot] <41898282+github-actions[bot]@users.noreply.github.com>